Move Aside para pasta components #6
